requirements.txt
inquirer==3.1.4
pandas==2.0.2
python-dotenv==1.0.0
Requests==2.31.0
